omnitrace-sample (#169)
- `omnitrace-sample` executable which executes sampling (no
instrumentation)
- fixes bug with OMPT ignoring value of `OMNITRACE_USE_OMPT`
- fixes some issues with sampling duration
- new `OMNITRACE_SAMPLING_INCLUDE_INLINES` configuration variable
- restricts process-sampling to 100 interrupts/sec when inheriting value
from `OMNITRACE_SAMPLING_FREQ`
- `OMNITRACE_PROCESS_SAMPLING_FREQ` still supports up to 1000
interrupts/sec
- fixes bug with colorized log not truly being disabled in all instances
- adds tests for `omnitrace-sample`
- adds tests for sampling duration
- settings ROCP_TOOL_LIB to libomnitrace-dl throws error
  - rocprofiler does not configure correctly when this is done
- Quiet numa_gotcha warnings
- Fixed some shadowed variables

+++ source/lib/omnitrace/library/components/rocprofiler.cpp
@@ -94,8 +94,7 @@ rocm_event::rocm_event(uint32_t _dev, uint32_t _thr, uint32_t _queue,
                 feature_values.emplace_back(rocm_feature_value{ p->data.result_int32 });
                 break;
             case ROCPROFILER_DATA_KIND_FLOAT:
-                feature_values.emplace_back(
-                    rocm_feature_value{ static_cast<double>(p->data.result_float) });
+                feature_values.emplace_back(rocm_feature_value{ p->data.result_float });
                 break;
             case ROCPROFILER_DATA_KIND_DOUBLE:
                 feature_values.emplace_back(rocm_feature_value{ p->data.result_double });
